Rename abi to metadata
