+++ src/workspace/mod.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with ink!.  If not, see <http://www.gnu.org/licenses/>.
 
-mod abi;
+mod metadata;
 mod manifest;
 mod profile;
